Run forge fmt on codebase

+++ src/property-checking/MerklePropertyChecker.sol
@@ -6,7 +6,6 @@ import {Clone} from "clones-with-immutable-args/Clone.sol";
 import {MerkleProof} from "@openzeppelin/contracts/utils/cryptography/MerkleProof.sol";
 
 contract MerklePropertyChecker is IPropertyChecker, Clone {
-
     // Immutable params
 
     /**
@@ -16,11 +15,11 @@ contract MerklePropertyChecker is IPropertyChecker, Clone {
         return bytes32(_getArgUint256(0));
     }
 
-    function hasProperties(uint256[] calldata ids, bytes calldata params) external pure returns(bool isAllowed) {
+    function hasProperties(uint256[] calldata ids, bytes calldata params) external pure returns (bool isAllowed) {
         isAllowed = true;
         bytes32 root = getMerkleRoot();
         (bytes[] memory proofList) = abi.decode(params, (bytes[]));
-        for (uint i; i < ids.length; i++) {
+        for (uint256 i; i < ids.length; i++) {
             bytes32[] memory proof = abi.decode(proofList[i], (bytes32[]));
             if (!MerkleProof.verify(proof, root, keccak256(abi.encodePacked(ids[i])))) {
                 return false;
